requirements.txt
cohere==4.32
weaviate-client==3.25.3
flask==2.3.2